pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "t212-exit-tax"
version = "0.1.0"
description = "Trading 212 Exit Tax calculator for Ireland"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "python-dotenv>=1.0.0",
  "rich>=13.0.0",
  "httpx>=0.27.0"
]

[project.scripts]
t212-exit-tax = "t212_exit_tax.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]